break out core game logic
